==> notebooks/3_feature_engineering.ipynb <==
Configuração Inicial

import findspark
findspark.init()
findspark.find()

from pyspark.sql import SparkSession
from pyspark.sql.functions import col, when, udf
from pyspark.sql.types import IntegerType
from pyspark.ml.feature import StringIndexer, OneHotEncoder, VectorAssembler, StandardScaler

# Inicialização da sessão Spark
spark = SparkSession\
        .builder\
        .appName("Projeto_AJP")\
        .master("local[*]")\
        .config("spark.executor.memory", "8g")\
        .config("spark.driver.memory", "8g")\
        .config("spark.driver.maxResultSize", "1g")\
        .config("spark.memory.offHeap.enabled", False)\
        .enableHiveSupport()\
        .getOrCreate()

Carregar Dados

# Carregamento dos dados pré-processados
flights = spark.read.parquet("../data/processed/flights_cleaned.parquet")

Criar Variável Alvo

# Criar coluna binária para indicar atraso na partida (1 se atraso > 0, caso contrário 0)
flights = flights.withColumn('IsDelayed', when(col('DepDelayMinutes') > 0, 1).otherwise(0))

Retirar Registos de Voos Cancelados ou Desviados

# Filtrar voos cancelados ou desviados
flights = flights.filter((col('Cancelled') == 0) & (col('Diverted') == 0))

# Remover colunas que não serão mais necessárias
flights = flights.drop('Cancelled', 'Diverted', 'DepDelayMinutes')

Indexação e Codificação

# Indexar colunas categóricas
categorical_cols = ['Airline', 'Origin', 'Dest']
indexers = [StringIndexer(inputCol=col, outputCol=col+"_Index") for col in categorical_cols]

# OneHotEncoding para colunas categóricas indexadas
encoders = [OneHotEncoder(inputCol=col+"_Index", outputCol=col+"_OHE") for col in categorical_cols]

# Construir o vetor de características
assembler = VectorAssembler(inputCols=[
    'Year', 'Month', 'DayofMonth', 'DayOfWeek', 'Distance',
    'CRSDepTime', 'Airline_OHE', 'Origin_OHE', 'Dest_OHE'], outputCol='features')

# Normalização das características
scaler = StandardScaler(inputCol='features', outputCol='scaledFeatures')

# Pipeline de transformação
from pyspark.ml import Pipeline
pipeline = Pipeline(stages=indexers + encoders + [assembler, scaler])

# Ajustar e transformar os dados
flights_transformed = pipeline.fit(flights).transform(flights)

Transformação do Dataframe final

# Selecionar apenas as colunas necessárias para o modelo
flights_transformed = flights_transformed.select('scaledFeatures', 'IsDelayed')

flights_transformed.count()

Salvar Dados

# Salvar os dados transformados em formato Parquet
flights_transformed.write.mode("overwrite").parquet("../data/processed/flights_features.parquet")

print("Feature Engineering completa e dados salvos.")

Encerrar Sessão Spark

spark.stop()
