==> notebooks/4_model_training.ipynb <==
Configurações Iniciais e Carregamento dos Dados

import findspark
findspark.init()
findspark.find()

from pyspark.sql import SparkSession
from pyspark.ml.classification import LogisticRegression, DecisionTreeClassifier, RandomForestClassifier, GBTClassifier
from pyspark.sql.functions import col

# Inicialização da sessão Spark
spark = SparkSession\
        .builder\
        .appName("Projeto_AJP")\
        .master("local[*]")\
        .config("spark.executor.memory", "8g")\
        .config("spark.driver.memory", "8g")\
        .config("spark.driver.maxResultSize", "1g")\
        .config("spark.memory.offHeap.enabled", False)\
        .enableHiveSupport()\
        .getOrCreate()

# Carregar os dados transformados
flights = spark.read.parquet("../data/processed/flights_features.parquet")

# Verificar os dados carregados
flights.show(5)

Separar os Dados em Treinamento e Teste

# Dividir os dados em 80% para treinamento e 20% para teste
train_data, test_data = flights.randomSplit([0.8, 0.2], seed=1234)

# Verificar as divisões
print("Train data count: ", train_data.count())
print("Test data count: ", test_data.count())

# Salvar dados teste para posterior avaliação
test_data.write.mode("overwrite").parquet("../data/processed/test_data.parquet")

Balanceamento dos Dados

# Contagem de exemplos em cada classe
class_counts = train_data.groupBy('IsDelayed').count().collect()
minority_class_count = min(class_counts, key=lambda x: x['count'])['count']
majority_class_count = max(class_counts, key=lambda x: x['count'])['count']

print(minority_class_count)

print(majority_class_count)

# Calcular a taxa de undersampling
undersample_ratio = minority_class_count / majority_class_count

# Aplicar undersampling à classe majoritária
undersampled_majority = train_data.filter(col('IsDelayed') == 0).sample(False, undersample_ratio, seed=1234)

# Combinar com a classe minoritária
balanced_train_data = train_data.filter(col('IsDelayed') == 1).unionAll(undersampled_majority)

print(balanced_train_data.count())

# Contagem de exemplos em cada classe
class_counts_balanced = balanced_train_data.groupBy('IsDelayed').count().collect()
minority_class_count_balanced = min(class_counts_balanced, key=lambda x: x['count'])['count']
majority_class_count_balanced = max(class_counts_balanced, key=lambda x: x['count'])['count']

# Calcular a taxa de undersampling
undersample_ratio_balanced = minority_class_count_balanced / majority_class_count_balanced

print(undersample_ratio_balanced)

Definição e Treinamento dos Modelos

##REGRESSÃO LOGISTICA

# Definir o modelo
lr = LogisticRegression(featuresCol='scaledFeatures', labelCol='IsDelayed')

# Treinar o modelo
lr_model = lr.fit(balanced_train_data)

# Gravar o modelo
lr_model.write().overwrite().save("../models/logistic_regression_model")

##ÁRVORE DE DECISÃO

dt = DecisionTreeClassifier(featuresCol='scaledFeatures', labelCol='IsDelayed')

dt_model = dt.fit(balanced_train_data)

dt_model.write().overwrite().save("../models/decision_tree_model")

##RANDOM FOREST

rf = RandomForestClassifier(featuresCol='scaledFeatures', labelCol='IsDelayed')

rf_model = rf.fit(balanced_train_data)

rf_model.write().overwrite().save("../models/random_forest_model")

Encerrar Sessão

spark.stop()
